% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{absolute_mean_correlation}
\alias{absolute_mean_correlation}
\title{Absolute Mean Correlation Coefficient}
\usage{
absolute_mean_correlation(cor_mat)
}
\arguments{
\item{cor_mat}{A correlation matrix.}
}
\value{
A numeric value of absoulte mean correlation coeffiecient
}
\description{
Calculates absolute mean correlation coefficient
\loadmathjax
}
\details{
The method calculates absolute mean correlation coeffiecient (MCC) by
solving a linear assignment problem. MCC is mathematically given as
\mjdeqn{
MCC(\mathbf K)=\frac{1}{p} \sup_{\mathbf P  \in \mathcal{P}}
tr(\mathbf P\, \abs(\mathbf K)),
}{ascii}
where \mjeqn{\mathcal{P}}{ascii} is a set of all possible
permutation matrices,
\mjeqn{ tr(\cdot )}{ascii} is the trace of a matrix, and
\mjeqn{\abs(\cdot)}{ascii}
denotes taking the absolute value of a matrix elementwise.
}
\examples{
d1 <- matrix(rnorm(300), ncol = 3)
d2 <- d1 + matrix(rnorm(300), ncol = 3)
cor_mat <- cor(d1, d2)
absolute_mean_correlation(cor_mat)
}
