% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVAE.R
\name{CVAE}
\alias{CVAE}
\title{Variational Autoencoder}
\usage{
CVAE(
  data,
  aux_data,
  latent_dim,
  hidden_units = c(128, 128, 128),
  validation_split = 0,
  activation = "leaky_relu",
  source_dist = "gaussian",
  error_dist = "gaussian",
  error_dist_sigma = 0.01,
  optimizer = NULL,
  lr_start = 0.01,
  lr_end = 1e-04,
  steps = 10000,
  seed = NULL,
  epochs,
  batch_size
)
}
\arguments{
\item{data}{A matrix with P columns and n rows}

\item{latent_dim}{A latent dimension for VAE}

\item{hidden_units}{K-dimensional vector giving the number of
hidden units for K layers in encoder and K layers in decoder.}

\item{validation_split}{Proportion of data used for validation}

\item{activation}{Activation function for the hidden layers.}

\item{source_dist}{Distribution for the latent source.
Either "gaussian" or "laplace".}

\item{error_dist}{Distribution for the model error.
Either "gaussian" or "laplace".}

\item{error_dist_sigma}{A standard deviation for error_dist.}

\item{optimizer}{A keras optimizer for the tensorflow model.
A default is Adam optimizer with polynomial decay.}

\item{lr_start}{A starting learning rate for the default optimizer.}

\item{lr_end}{A ending learning rate for the default optimizer.}

\item{steps}{A number of learning steps between lr_start and lr_end
for the default optimizer.}

\item{seed}{Seed for the tensorflow model. Should be used instead of
set.seed(seed).}

\item{epochs}{A number of epochs for training.}

\item{batch_size}{A batch size for training.}
}
\value{
An object of class VAE.
\item{IC_unscaled}{Unscaled latent components.}
\item{IC}{The latent component with
zero mean and unit variance.}
\item{data_dim}{The dimension of the original data.}
\item{metrics}{Metrics of the training for each epoch.}
\item{sample_size}{Sample size of
the data.}
\item{call_params}{The params for
the original VAE method call.}
\item{encoder}{The trained encoder.}
\item{decoder}{The trained decoder.}
\item{IC_means}{The means of the
unscaled latent components.}
\item{IC_means}{The standard devivations
of the unscaled latent components.}
\item{call}{The of how the method was called.}
\item{DNAME}{The of the original data.}
}
\description{
Trains a variational autoencoder (VAE) using the input data.
}
\details{
The method constructs and trains a variational autoencoder (VAE)
\insertCite{kingma2013auto}{NonlinearBSS} based on the given parameters.
VAE is composed of an encoder and a decoder. The encoder transforms the
original data into a latent representation. The decoder aims to
transform the latent representation back to the original data.
The variational approximation is obtained by using a reparametrization
trick to sample a new value using the mean and the standard deviation
given by the encoder.
}
\examples{
p <- 3
n <- 1000
latent_data <- matrix(rnorm(p * n), ncol = p, nrow = n)
mixed_data <- mix_data(latent_data, 2, "elu")
res <- VAE(mixed_data, p, hidden_units = c(128, 64, 32), 
lr_start = 0.0001, epochs = 1, batch_size = 64)
}
\references{
\insertAllCited{}
}
