% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{generate_nonstationary_spatio_temporal_data_by_segments}
\alias{generate_nonstationary_spatio_temporal_data_by_segments}
\title{Generate Nonstationary Spatio-Temporal Data by Segments}
\usage{
generate_nonstationary_spatio_temporal_data_by_segments(
  n_time,
  n_spat,
  p,
  n_segments_space,
  n_segments_time,
  coords_time
)
}
\arguments{
\item{n_time}{Number of time points.}

\item{n_spat}{Number of spatial points.}

\item{p}{A dimension of the data.}

\item{n_segments_space}{Number of spatial segments.}

\item{n_segments_time}{Number of temporal segments.}

\item{coords_time}{A 3 x n (x, y, time) matrix containing the 
spatio-temporal coordinates.}
}
\value{
An object with the following properties
\item{data}{A p x n matrix containing the generated data.}
\item{labels}{A vector giving numerical labels for each data point.
The data points with same label belong in the same segment.}
\item{coords}{A 3 x n matrix containing the spatio-temporal coordinates.}
}
\description{
Generates nonstationary Gaussian spatio-temporal data with
changing mean and variance by segments. \loadmathjax
}
\details{
The method chooses uniformly \code{n_segments_space} center points from 
the provided spatial locations. The locations are divided into segments 
based on the lowest distance to any of the center points. Similarly 
\code{n_segments_time} center points are selected from the provided 
temporal points and the data are divided into temporal segments.
The data is generated from normal distribution by sampling unique variance 
from distribution \mjeqn{Unif(0.1, 5)}{ascii} for each segment and mean from
distribution \mjeqn{Unif(-1, 1)}{ascii}.
}
\examples{
coords_time <- cbind(rep(runif(50), 100), rep(runif(50), 100), rep(1:100, each=50))
data_obj <- generate_nonstationary_spatio_temporal_data_by_segments(100, 50, 3, 5, 10, coords_time)
}
