% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iVAE_coords.R
\name{iVAE_coords}
\alias{iVAE_coords}
\title{Radial Basis Function Based Spatio Temporal Identifiable Variational Autoencoder}
\usage{
iVAE_coords(data, locations, latent_dim, epochs, batch_size, ...)
}
\arguments{
\item{data}{A matrix with P columns and N rows containing the observed data.}

\item{locations}{A matrix containing spatial or spatio-temporal locations.}

\item{latent_dim}{A latent dimension for iVAE.}

\item{epochs}{A number of epochs to train the model.}

\item{batch_size}{A batch size.}

\item{...}{
  Arguments passed on to \code{\link[=iVAE]{iVAE}}
  \describe{
    \item{\code{hidden_units}}{K-dimensional vector giving the number of
hidden units for K layers in encoder and K layers in decoder.}
    \item{\code{aux_hidden_units}}{K-dimensional vector giving the number of
hidden units for K layers in auxiliary function.}
    \item{\code{validation_split}}{Proportion of data used for validation}
    \item{\code{activation}}{Activation function for the hidden layers.}
    \item{\code{source_dist}}{Distribution for the latent source.
Either "gaussian" or "laplace".}
    \item{\code{error_dist}}{Distribution for the model error.
Either "gaussian" or "laplace".}
    \item{\code{error_dist_sigma}}{A standard deviation for error_dist.}
    \item{\code{optimizer}}{A keras optimizer for the tensorflow model.
A default is Adam optimizer with polynomial decay.}
    \item{\code{lr_start}}{A starting learning rate for the default optimizer.}
    \item{\code{lr_end}}{A ending learning rate for the default optimizer.}
    \item{\code{steps}}{A number of learning steps between lr_start and lr_end
for the default optimizer.}
    \item{\code{seed}}{Seed for the tensorflow model. Should be used instead of
set.seed(seed).}
    \item{\code{get_elbo}}{Logical. If TRUE, the model returns the final ELBO value. Default is FALSE.}
  }}
}
\value{
An object of class iVAEcoords, inherits from class iVAE.
Additionally, the object has the following properties:
\item{location_mins}{A vector containing the minimum
locations of each dimension.}
\item{location_maxs}{A vector containing the maximum
locations of each dimension.}
\item{spatial_dim}{The spatial dimension.}
For more details, see \code{\link{iVAE}}.
}
\description{
Trains an identifiable variational autoencoder (iVAE) which uses
spatial and temporal radial basis functions as auxiliary data.
}
\details{
The method uses the min-max normalized coordinates directly as the auxiliary
data.

After the forming the radial basis functions, the method
calls the function \code{iVAE} using the created auxiliary variables.
}
\examples{
p <- 3
n_time <- 100
n_spat <- 50
coords_time <- cbind(
    rep(runif(n_spat), n_time), rep(runif(n_spat), n_time),
    rep(1:n_time, each = n_spat)
)
data_obj <- generate_nonstationary_spatio_temporal_data_by_segments(
    n_time,
    n_spat, p, 5, 10, coords_time
)
latent_data <- data_obj$data
# Generate artificial observed data by applying a nonlinear mixture
obs_data <- mix_data(latent_data, 2)
cor(obs_data, latent_data)

# For better peformance, increase the number of epochs.
resiVAE <- iVAE_coords(obs_data, coords_time, p,
    epochs = 10, batch_size = 64
)
cormat <- cor(resiVAE$IC, latent_data)
cormat
absolute_mean_correlation(cormat)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{iVAE}}
\code{\link{generate_nonstationary_spatio_temporal_data_by_segments}}
}
\author{
Mika Sipilä
}
