% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iVAE_radial_spatial.R
\name{iVAE_radial_spatial}
\alias{iVAE_radial_spatial}
\title{Radial Basis Function Based Spatial Identifiable Variational Autoencoder}
\usage{
iVAE_radial_spatial(
  data,
  locations,
  latent_dim,
  num_basis = c(2, 9),
  kernel = "gaussian",
  epochs,
  batch_size,
  ...
)
}
\arguments{
\item{data}{A matrix with P columns and N rows containing the observed data.}

\item{locations}{A matrix with spatial locations.}

\item{latent_dim}{A latent dimension for iVAE.}

\item{num_basis}{A vector containing the number of basis functions for
each resolution. Default value is (2, 9).}

\item{kernel}{A kernel function to be used to form the radial basis
functions. Either \code{"gaussian"} (default) or \code{"wendland"}.}

\item{epochs}{A number of epochs to train the model.}

\item{batch_size}{A batch size.}

\item{...}{
  Arguments passed on to \code{\link[=iVAE]{iVAE}}
  \describe{
    \item{\code{hidden_units}}{K-dimensional vector giving the number of
hidden units for K layers in encoder and K layers in decoder.}
    \item{\code{aux_hidden_units}}{K-dimensional vector giving the number of
hidden units for K layers in auxiliary function.}
    \item{\code{validation_split}}{Proportion of data used for validation}
    \item{\code{activation}}{Activation function for the hidden layers.}
    \item{\code{source_dist}}{Distribution for the latent source.
Either "gaussian" or "laplace".}
    \item{\code{error_dist}}{Distribution for the model error.
Either "gaussian" or "laplace".}
    \item{\code{error_dist_sigma}}{A standard deviation for error_dist.}
    \item{\code{optimizer}}{A keras optimizer for the tensorflow model.
A default is Adam optimizer with polynomial decay.}
    \item{\code{lr_start}}{A starting learning rate for the default optimizer.}
    \item{\code{lr_end}}{A ending learning rate for the default optimizer.}
    \item{\code{steps}}{A number of learning steps between lr_start and lr_end
for the default optimizer.}
    \item{\code{seed}}{Seed for the tensorflow model. Should be used instead of
set.seed(seed).}
    \item{\code{get_elbo}}{Logical. If TRUE, the model returns the final ELBO value. Default is FALSE.}
  }}
}
\value{
An object of class iVAEradial, inherits from class iVAE.
Additionally, the object has the following properties:
\code{min_coords} A vector of minimum coordinates.
\code{max_coords} A vector of maximum coordinates.
\code{num_basis} Same as \code{num_basis} in the function call.
\code{spatial_dim} The dimension the given locations.
For more details, see \code{\link{iVAE}}.
}
\description{
Trains an identifiable variational autoencoder (iVAE) using the input data
and spatial radial basis functions as auxiliary data.
}
\details{
The method creates the auxiliary data as radial basis functions based on
the given input parameters.
The vector \code{num_basis} defines the resolution levels used to create the
spatial radial basis functions. For example \code{num_basis = c(2, 9)}
creates the radial basis functions with resolution levels 2 and 9. The
radial basis functions are created identically as in
\insertCite{chen2020deepkriging}{NonlinearBSS}.

After the forming the radial basis functions, the method
calls the function \code{iVAE} using the created auxiliary variables.
}
\examples{
n <- 1000
coords <- matrix(runif(1000 * 2, 0, 1), ncol = 2)
p <- 3
# Generate artificial latent data
latent_data <- generate_nonstationary_spatial_data_by_segments(
    n, p,
    coords, 10
)$data
# Generate artificial observed data by applying a nonlinear mixture
obs_data <- mix_data(latent_data, 2)
cor(obs_data, latent_data)

# For better peformance, increase the number of epochs.
resiVAE <- iVAE_radial_spatial(obs_data, coords, 3,
    epochs = 10, batch_size = 64
)
cormat <- cor(resiVAE$IC, latent_data)
cormat
absolute_mean_correlation(cormat)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{iVAE}}
\code{\link{generate_nonstationary_spatial_data_by_segments}}
}
\author{
Mika Sipilä
}
