% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{laplace_log_pdf}
\alias{laplace_log_pdf}
\title{Calculates a logarithm of multivariate laplace probability density function using tensorflow.}
\usage{
laplace_log_pdf(x, loc, scale, reduce = TRUE)
}
\arguments{
\item{x}{A P-dimensional vector}

\item{mu}{A P-dimensional location vector}

\item{v}{A P-dimensional scale vector for diagonal scale matrix}
}
\value{
A logarithm of multivariate laplace pdf.
}
\description{
Calculates a logarithm of multivariate laplace probability density function using tensorflow.
}
