% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iVAE.R
\name{load_with_tf}
\alias{load_with_tf}
\title{Load iVAE Object with Trained Tensorflow Models}
\usage{
load_with_tf(file)
}
\arguments{
\item{file}{Filename of the saved \code{iVAE} object.}
}
\value{
A loaded \code{iVAE} object with restored trained Tensorflow models.
}
\description{
Loads \code{iVAE} object including the trained
tensorflow models.
}
\details{
The function loads iVAE object correctly by loading also the trained
tensorflow models.
}
\examples{
p <- 3
n_segments <- 10
n_per_segment <- 100
n <- n_segments * n_per_segment
latent_data <- matrix(NA, ncol = p, nrow = n)
aux_data <- matrix(0, ncol = n_segments, nrow = n)
# Create artificial data with variance and mean varying over the segments.
for (i in 1:p) {
  for (seg in 1:n_segments) {
    start_ind <- (seg - 1) * n_per_segment + 1
    end_ind <- seg * n_per_segment
    latent_data[start_ind:end_ind, i] <- rnorm(
      n_per_segment,
      runif(1, -5, 5), runif(1, 0.1, 5)
    )
  }
}
mixed_data <- mix_data(latent_data, 2, "elu")
# Increase the number of epochs to obtain better performance.
res <- iVAE(mixed_data, aux_data, 3, epochs = 10, batch_size = 64)
save_with_tf(res, "res_dir", "res_obj.RData")
loaded_obj <- load_with_tf("res_obj.RData")
new_ICs <- matrix(rnorm(p * 2), nrow = 2)
pred_obs <- predict(loaded_obj, new_ICs, IC_to_data = TRUE)
}
\seealso{
\code{\link{save_with_tf}}
}
\author{
Mika Sipilä
}
