% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mix_data_over_determinated}
\alias{mix_data_over_determinated}
\title{Create Overdeterminated Mixed Data}
\usage{
mix_data_over_determinated(data, dim_mixed, n_layers = 1, nonlinearity = "elu")
}
\arguments{
\item{data}{A matrix containing the original data.}

\item{dim_mixed}{The dimension of the mixed data.}

\item{n_layers}{The number of mixing layers.}

\item{nonlinearity}{The function for applying nonlinearity
between the mixing layers.}
}
\value{
A matrix containing the mixed data.
}
\description{
Creates overdeterminated mixed data based on the input data.
\loadmathjax
}
\details{
The method mixes the input data by applying mixing matrices and nonlinear
activation function based on the input parameters. 
Let \mjeqn{\omega_i}{ascii} be
the activation function of \mjeqn{i}{ascii}th layer and 
\mjeqn{\mathbf B_i}{ascii} be the
normalized mixing matrix of \mjeqn{i}{ascii}th layer.
Then, the mixing function \mjeqn{\mathbf f_L}{ascii} 
is defined as
\mjdeqn{
    \mathbf f_L(\mathbf z) = \begin{cases}
        \omega_L(\mathbf B_L \mathbf z),\quad L = 1, \\
        \omega_L(\mathbf B_L \mathbf f_{L-1}(\mathbf z)),
        \quad L \in \{2,3,\dots\}.
    \end{cases}
}{ascii}
The function has a linear activation 
\mjeqn{\omega_L(x)=x}{ascii} for the last layer,
which means that mixing function 
\mjeqn{\mathbf f_1}{ascii} with one layer corresponds
to a linear mixing. If more than one layers are applied, the other layers
use the activation function given by the parameter \code{nonlinearity}.
The options for \code{nonlinearity} are:

\code{nonlinarity="elu"}:
\mjdeqn{
\omega_i(x)=\begin{cases}
    x,\quad x \geq 0, \\
    exp(x) - 1,\quad x < 0,
\end{cases}
}{ascii}

\code{nonlinarity="xtanh"}:
\mjdeqn{
\omega_i(x)= tanh(x) + 0.1 * x
}{ascii}

\code{nonlinarity="lrelu"}:
\mjdeqn{
\omega_i(x)=\begin{cases}
    x,\quad x \geq 0, \\
    0.2 * x, x < 0.
\end{cases}
}{ascii}
}
\examples{
data <- matrix(rnorm(300), ncol = 3)
mixed_data <- mix_data_over_determinated(data, 5, 3)
}
