\name{NonlinearBSS-package}
\alias{NonlinearBSS-package}
\alias{NonlinearBSS}
\docType{package}
\title{
\packageTitle{NonlinearBSS}
}
\description{
\packageDescription{NonlinearBSS}
}
\details{
The package contains tools to perform nonlinear blind source separation. The methods include regular variational autoencoder \code{VAE}, some identifiable variational autoencoder (iVAE) based methods including general version \code{iVAE} and a version for spatial data \code{iVAE_spatial}. In addition, a time contrastive learning (TCL) algorithm and various methods for generating data have been implemented.
}
\author{
Authors: Mika Sipilä, Klaus Nordhausen, Sara Taskinen

Maintainer: Mika Sipilä
}
\references{


\cite{I. Khemakhem, D. Kingma, R. Monti, A. Hyvarinen, Variational au-
toencoders and nonlinear ICA: A unifying framework, in: International
Conference on Artificial Intelligence and Statistics, PMLR, 2020, pp.
2207–2217. <doi:https://doi.org/10.48550/arXiv.1907.04809>.}
}

\keyword{ package }