% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{norm_log_pdf}
\alias{norm_log_pdf}
\title{Calculates a logarithm of multivariate normal probability density function using tensorflow.}
\usage{
norm_log_pdf(x, mu, v, reduce = TRUE)
}
\arguments{
\item{x}{A P-dimensional vector}

\item{mu}{A P-dimensional mean vector}

\item{v}{A P-dimensional variance vector for diagonal covariance}
}
\value{
A logarithm of multivariate normal pdf.
}
\description{
Calculates a logarithm of multivariate normal probability density function using tensorflow.
}
