% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iVAE.R
\name{plot.iVAE}
\alias{plot.iVAE}
\title{Plotting an Object of Class iVAE}
\usage{
\method{plot}{iVAE}(
  x,
  IC_inds = 1:x$call_params$latent_dim,
  sample_inds = 1:dim(x$IC)[1],
  unscaled = FALSE,
  type = "l",
  xlab = "",
  ylabs = c(),
  colors = c(),
  oma = c(1, 1, 0, 0),
  mar = c(2, 2, 1, 1),
  ...
)
}
\arguments{
\item{IC_inds}{A vector providing indices of ICs which are plotted.}

\item{sample_inds}{A vector providing sample indices which are plotted.}

\item{unscaled}{A boolean determining if the unscaled ICs are plotted
or not.}

\item{type}{linetype for plot.default function.}

\item{xlab}{xlab for the bottom most plot.}

\item{ylabs}{A vector providing ylab params for the plots.}

\item{colors}{A vector providing colors for the plots.}

\item{oma}{A \code{oma} parameter for the function \code{par}.
See \code{\link[base]{par}}}

\item{mar}{A \code{mar} parameter for the function \code{par}.
See \code{\link[base]{par}}}

\item{...}{Further parameters for all plot.default methods.}

\item{obj}{Object of class \code{iVAE}}
}
\value{
No return value.
}
\description{
\code{plot} method for the class \code{iVAE}.
}
\details{
Plots the components provided by the parameter \code{IC_inds}. A sample
subset can be provided by parameter \code{sample_inds}.
}
\examples{
p <- 3
n_segments <- 10
n_per_segment <- 100
n <- n_segments * n_per_segment
latent_data <- matrix(NA, ncol = p, nrow = n)
aux_data <- matrix(NA, ncol = n_segments, nrow = n)
# Create artificial data with variance and mean varying over the segments.
for (i in 1:p) {
  for (seg in 1:n_segments) {
    start_ind <- (seg - 1) * n_per_segment + 1
    end_ind <- seg * n_per_segment
    latent_data[start_ind:end_ind, i] <- rnorm(
      n_per_segment,
      runif(1, -5, 5), runif(1, 0.1, 5)
    )
  }
}
mixed_data <- mix_data(latent_data, 2, "elu")
res <- iVAE(mixed_data, aux_data, 3, epochs = 300, batch_size = 64)
plot(res)
}
\seealso{
\code{\link{iVAE}}
}
\author{
Mika Sipilä
}
