% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iVAE.R
\name{predict.iVAE}
\alias{predict.iVAE}
\title{Predict Method for an Object of Class iVAE}
\usage{
\method{predict}{iVAE}(object, newdata, aux_data = NULL, IC_to_data = FALSE, ...)
}
\arguments{
\item{object}{Object of class \code{iVAE}}

\item{newdata}{A matrix containing the new data. It can contain
new observations or new latent component. Choose the parameter
\code{IC_to_data} accordingly.}

\item{aux_data}{A matrix containing auxiliary data for new observations.
Should not be provided if \code{IC_to_data} is \code{TRUE}.}

\item{IC_to_data}{A boolean defining if the prediction is done from
latent components to observations or from observations to latent
components. If \code{IC_to_data = TRUE}, \code{newdata} should contain
new latent components. Otherwise it should contain the new observations.}

\item{...}{Further parameters for \code{predict.default}.}
}
\value{
A matrix containing the predictions.
}
\description{
\code{predict} method for the class \code{iVAE}.
}
\details{
Summarizes the object of class \code{iVAE}.
}
\examples{
p <- 3
n_segments <- 10
n_per_segment <- 100
n <- n_segments * n_per_segment
latent_data <- matrix(NA, ncol = p, nrow = n)
aux_data <- matrix(NA, ncol = n_segments, nrow = n)
# Create artificial data with variance and mean varying over the segments.
for (i in 1:p) {
  for (seg in 1:n_segments) {
    start_ind <- (seg - 1) * n_per_segment + 1
    end_ind <- seg * n_per_segment
    latent_data[start_ind:end_ind, i] <- rnorm(
      n_per_segment,
      runif(1, -5, 5), runif(1, 0.1, 5)
    )
  }
}
mixed_data <- mix_data(latent_data, 2, "elu")
res <- iVAE(mixed_data, aux_data, 3, epochs = 300, batch_size = 64)
new_data <- matrix(rnorm(p * 2), nrow = 2)
new_aux_data <- rbind(c(1, rep(0, n_segments - 1)), c(1, rep(0, n_segments - 1)))
pred_ICs <- predict(res, new_data, new_aux_data)
new_ICs <- matrix(rnorm(p * 2), nrow = 2)
pred_obs <- predict(res, new_ICs, IC_to_data = TRUE)
}
\seealso{
\code{\link{iVAE}}
}
