% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_coords_to_IC.R
\name{predict_coords_to_IC}
\alias{predict_coords_to_IC}
\title{Predict the Latent Components from Coordinates}
\usage{
predict_coords_to_IC(
  object,
  new_spatial_locations,
  new_time_points,
  new_elevation = NULL,
  get_var = FALSE
)
}
\arguments{
\item{object}{An object of class \code{iVAE_radial_st}}

\item{new_spatial_locations}{A matrix containing the new spatial locations.}

\item{new_time_points}{A vector containing the new temporal points.}

\item{new_elevation}{An optional vector containing elevation for new data.}

\item{get_var}{A boolean. If \code{TRUE}, the method returns the
spatio-temporal variance function estimate.
Otherwise returns the trend function estimate.}
}
\value{
A matrix containing the predicted variance or trend function for
the latent components.
}
\description{
Predicts the latent components from the coordinates using
an object of class \code{iVAE_radial_st}
}
\details{
The method uses trained spatio-temporal identifiable variational autoencoder
to predict the latent components directly using new coordinates provided
in the function input.
}
\examples{
p <- 3
n_time <- 100
n_spat <- 50
coords_time <- cbind(
    rep(runif(n_spat), n_time), rep(runif(n_spat), n_time),
    rep(1:n_time, each = n_spat)
)
data_obj <- generate_nonstationary_spatio_temporal_data_by_segments(
    n_time,
    n_spat, p, 5, 10, coords_time
)
latent_data <- data_obj$data
# Generate artificial observed data by applying a nonlinear mixture
obs_data <- mix_data(latent_data, 2)

# For better peformance, increase the number of epochs.
resiVAE_st <- iVAE_radial_spatio_temporal(obs_data, coords_time[, 1:2], coords_time[, 3], p,
    epochs = 10, batch_size = 64
)

new_spatial_coords <- cbind(runif(10), runif(10))
new_time_points <- rep(1, 10)

trend_estimate <- predict_coords_to_IC(
    resiVAE_st, new_spatial_coords,
    new_time_points
)
variance_estimate <- predict_coords_to_IC(resiVAE_st, new_spatial_coords,
    new_time_points,
    get_var = TRUE
)
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{iVAE_radial_spatio_temporal}}
}
\author{
Mika Sipilä
}
