% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_coords_to_IC_ar_mean.R
\name{predict_coords_to_IC_ar}
\alias{predict_coords_to_IC_ar}
\title{Predict Latent Independent Components for New Spatial-Temporal Locations}
\usage{
predict_coords_to_IC_ar(
  object,
  last_spatial_locations,
  last_time_points,
  last_elevations = NULL,
  new_spatial_locations,
  new_time_points,
  new_elevation = NULL,
  get_trend = FALSE,
  get_var = FALSE,
  get_ar_coefs = FALSE,
  unscaled = FALSE
)
}
\arguments{
\item{object}{An object of class `iVAEradial_st`.}

\item{last_spatial_locations}{A matrix of spatial coordinates corresponding to the last known time points.}

\item{last_time_points}{A vector of time points corresponding to `last_spatial_locations`.}

\item{last_elevations}{(Optional) A vector of elevation values corresponding to `last_spatial_locations`.}

\item{new_spatial_locations}{A matrix of new spatial coordinates where predictions are required.}

\item{new_time_points}{A vector of time points corresponding to `new_spatial_locations`.}

\item{new_elevation}{(Optional) A vector of elevation values corresponding to `new_spatial_locations`.}

\item{get_trend}{Logical; if `TRUE`, return the predicted trend functions of the ICs.}

\item{get_var}{Logical; if `TRUE`, return the predicted variance of the ICs.}

\item{get_ar_coefs}{Logical; if `TRUE`, return the predicted autoregressive (AR) coefficients of the ICs.}

\item{unscaled}{Logical; if `TRUE`, return unscaled predictions (raw ICs before standardization).}
}
\value{
A list containing:
  - `preds`: A matrix of predicted ICs.
  - `coords_time`: A matrix of corresponding spatial and temporal coordinates (with optional elevation).
  - `trends`: A matrix of predicted mean of the ICs if `get_trend = TRUE`, otherwise `NULL`.
  - `vars`: A matrix of predicted variances of the ICs if `get_var = TRUE`, otherwise `NULL`.
  - `ar_coefs`: A matrix of predicted autoregressive coefficients of ICs if `get_ar_coefs = TRUE`, otherwise `NULL`.
}
\description{
This function predicts the latent independent components (ICs) for new spatial-temporal locations 
based on the fitted autoregressive identifiable variational autoencoder (iVAEar).
}
\details{
This function utilizes an autoregressive iVAE model to predict 
latent independent components (ICs) at new spatial-temporal locations.
The method relies on a learned trend function \mjeqn{\mu(\mathbf{s}, t)}{ascii} 
and an autoregressive structure to propagate ICs over time.

Specifically, the model follows an autoregressive process of order \mjeqn{R}{ascii}:

\mjeqn{\mathbf{z}_t = \mu(\mathbf{s}, t) + \sum_{r=1}^{R} \gamma_r(\mathbf{s}, t) (\mathbf{z}_{t-r} - \mu(\mathbf{s}_t, t)) + \omega_t}{ascii}

where \mjeqn{\gamma_r(\mathbf{s}_t, t)}{ascii} are spatially varying autoregressive coefficients and 
\mjeqn{\omega_t}{ascii} represents stochastic noise.

Given new spatial and temporal coordinates, the function reconstructs ICs by 
sequentially applying the learned trend function and autoregressive model.
}
\author{
Mika Sipilä
}
