% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_coords_to_IC_ar_mean.R
\name{predict_coords_to_IC_ar_train_data}
\alias{predict_coords_to_IC_ar_train_data}
\title{Predict Latent Independent Components for Training Data}
\usage{
predict_coords_to_IC_ar_train_data(
  object,
  get_var = FALSE,
  get_ar_coefs = FALSE,
  unscaled = FALSE
)
}
\arguments{
\item{object}{An object of class `iVAEradial_st`.}

\item{get_var}{Logical; if `TRUE`, return the predicted variance of the ICs instead of the IC predictions.}

\item{get_ar_coefs}{Logical; if `TRUE`, return the autoregressive (AR) coefficients instead of IC predictions.}

\item{unscaled}{Logical; if `TRUE`, return unscaled predictions (raw ICs before standardization).}

\item{get_trend}{Logical; if `TRUE`, return the predicted trend functions of the ICs.}
}
\value{
A list containing:
  - `preds`: A matrix of predicted ICs.
  - `trends`: A matrix of predicted mean of the ICs if `get_trend = TRUE`, otherwise `NULL`.
  - `vars`: A matrix of predicted variances of the ICs if `get_var = TRUE`, otherwise `NULL`.
  - `ar_coefs`: A matrix of predicted autoregressive coefficients of ICs if `get_ar_coefs = TRUE`, otherwise `NULL`.
}
\description{
This function predicts the latent independent components (ICs) for the training data locations 
using the fitted autoregressive identifiable variational autoencoder (iVAEar).
}
\details{
This function utilizes an autoregressive iVAE model to predict 
latent independent components (ICs) at new spatial-temporal locations.
The method relies on a learned trend function \mjqen{\mu(\mathbf{s}, t)}{ascii} 
and an autoregressive structure to propagate ICs over time.

Specifically, the model follows an autoregressive process of order \mjqen{R}{ascii}:

\mjeqn{\mathbf{z}_t = \mu(\mathbf{s}, t) + \sum_{r=1}^{R} \gamma_r(\mathbf{s}, t) (\mathbf{z}_{t-r} - \mu(\mathbf{s}_t, t)) + \omega_t}{ascii}

where \mjeqn{\gamma_r(\mathbf{s}_t, t)}{ascii} are spatially varying autoregressive coefficients and 
\mjeqn{\omega_t}{ascii} represents stochastic noise.

Given new spatial and temporal coordinates, the function reconstructs ICs by 
sequentially applying the learned trend function and autoregressive model.
}
\author{
Mika Sipilä
}
