% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iVAE.R
\name{print.iVAE}
\alias{print.iVAE}
\title{Printing an Object of Class iVAE}
\usage{
\method{print}{iVAE}(x, ...)
}
\arguments{
\item{x}{Object of class \code{iVAE}}

\item{...}{Further parameters for \code{print.default}.}
}
\value{
No return value.
}
\description{
\code{print} method for the class \code{iVAE}.
}
\details{
Prints details about the object of class \code{iVAE}.
}
\examples{
p <- 3
n_segments <- 10
n_per_segment <- 100
n <- n_segments * n_per_segment
latent_data <- matrix(NA, ncol = p, nrow = n)
aux_data <- matrix(NA, ncol = n_segments, nrow = n)
# Create artificial data with variance and mean varying over the segments.
for (i in 1:p) {
  for (seg in 1:n_segments) {
    start_ind <- (seg - 1) * n_per_segment + 1
    end_ind <- seg * n_per_segment
    latent_data[start_ind:end_ind, i] <- rnorm(
      n_per_segment,
      runif(1, -5, 5), runif(1, 0.1, 5)
    )
  }
}
mixed_data <- mix_data(latent_data, 2, "elu")
res <- iVAE(mixed_data, aux_data, 3, epochs = 300, batch_size = 64)
print(res)
}
\seealso{
\code{\link{iVAE}}
}
