% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iVAE.R
\name{save_with_tf}
\alias{save_with_tf}
\title{Save iVAE Object with Trained Tensorflow Models}
\usage{
save_with_tf(object, tf_model_dir, file, ...)
}
\arguments{
\item{object}{Object of class \code{iVAE}.}

\item{tf_model_dir}{Directory, where the tensorflow.
models should be saved.}

\item{file}{Filename for saving the \code{iVAE} object.}

\item{...}{
  Arguments passed on to \code{\link[base:save]{base::save}}
  \describe{
    \item{\code{ascii}}{if \code{TRUE}, an ASCII representation of the data is
    written.  The default value of \code{ascii} is \code{FALSE} which
    leads to a binary file being written.  If \code{NA} and
    \code{version >= 2}, a different ASCII representation is used which
    writes double/complex numbers as binary fractions.}
    \item{\code{version}}{the workspace format version to use.  \code{NULL}
    specifies the current default format (3).  Version 1 was the default
    from \R 0.99.0 to \R 1.3.1 and version 2 from \R 1.4.0 to 3.5.0.
    Version 3 is supported from \R 3.5.0.}
    \item{\code{envir}}{environment to search for objects to be saved.}
    \item{\code{compress}}{logical or character string specifying whether saving
    to a named file is to use compression.  \code{TRUE} corresponds to
    \command{gzip} compression, and character strings \code{"gzip"},
    \code{"bzip2"} or \code{"xz"} specify the type of
    compression.  Ignored when \code{file} is a connection and
    for workspace format version 1.}
    \item{\code{compression_level}}{integer: the level of compression to be
    used.  Defaults to \code{6} for \command{gzip} compression and to
    \code{9} for \command{bzip2} or \command{xz} compression.  See the
    help for \code{\link[base]{file}} for possible values and their merits.}
    \item{\code{eval.promises}}{logical: should objects which are promises be
    forced before saving?}
    \item{\code{precheck}}{logical: should the existence of the objects be
    checked before starting to save (and in particular before opening
    the file/connection)?  Does not apply to version 1 saves.}
  }}
}
\value{
No return value.
}
\description{
Saves \code{iVAE} object including the trained
tensorflow models.
}
\details{
The function saves iVAE object correctly by saving also the trained
tensorflow models. In a new session, the trained model is available
by loading the \code{iVAE} object using the function
\code{\link{load_with_tf}}.
}
\examples{
p <- 3
n_segments <- 10
n_per_segment <- 100
n <- n_segments * n_per_segment
latent_data <- matrix(NA, ncol = p, nrow = n)
aux_data <- matrix(NA, ncol = n_segments, nrow = n)
# Create artificial data with variance and mean varying over the segments.
for (i in 1:p) {
  for (seg in 1:n_segments) {
    start_ind <- (seg - 1) * n_per_segment + 1
    end_ind <- seg * n_per_segment
    latent_data[start_ind:end_ind, i] <- rnorm(
      n_per_segment,
      runif(1, -5, 5), runif(1, 0.1, 5)
    )
  }
}
mixed_data <- mix_data(latent_data, 2, "elu")
res <- iVAE(mixed_data, aux_data, 3, epochs = 300, batch_size = 64)
save_with_tf(res, "res_dir", "res_obj.RData")
loaded_obj <- load_with_tf("res_obj.RData")
new_ICs <- matrix(rnorm(p * 2), nrow = 2)
pred_obs <- predict(loaded_obj, new_ICs, IC_to_data = TRUE)
}
\seealso{
\code{\link{load_with_tf}}
}
