% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{absolute_mean_correlation}
\alias{absolute_mean_correlation}
\title{Absolute Mean Correlation Coefficient}
\usage{
absolute_mean_correlation(cor_mat)
}
\arguments{
\item{cor_mat}{A correlation matrix.}
}
\value{
A numeric value of absoulte mean correlation coeffiecient
}
\description{
Calculates absolute mean correlation coefficient
}
\details{
The method calculates absolute mean correlation coeffiecient (MCC) by
solving a linear assignment problem. MCC is mathematically given as
\deqn{
\text{MCC}(\mathbf K)=\frac{1}{p} \sup_{\mathbf P  \in \mathcal{P}}
\text{tr}(\mathbf P\, \abs(\mathbf K)),
}
where \eqn{\mathcal{P}} is a set of all possible permutation matrices,
\eqn{\text{tr}(\cdot )} is the trace of a matrix, and \eqn{\abs(\cdot)}
denotes taking the absolute value of a matrix elementwise.
}
\examples{
d1 <- matrix(rnorm(300), ncol = 3)
d2 <- d1 + matrix(rnorm(300), ncol = 3)
cor_mat <- cor(d1, d2)
absolute_mean_correlation(cor_mat)
}
