% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{generate_nonstationary_spatial_data_by_segments}
\alias{generate_nonstationary_spatial_data_by_segments}
\title{Generate Nonstationary Spatial Data by Segments}
\usage{
generate_nonstationary_spatial_data_by_segments(
  n,
  p,
  coords,
  n_segments,
  random_mean = TRUE
)
}
\arguments{
\item{n}{A sample size.}

\item{p}{A dimension of the data.}

\item{coords}{A 2 x n matrix containing the spatial coordinates.}

\item{n_segments}{The number of spatial segments. Each segment
has their own mean and variance}

\item{random_mean}{A boolean determining if the constant zero mean
is used or if mean is random randomly sampled from
\mjeqn{Unif(-5, 5)}{ascii}
for each segment.}
}
\value{
An object with the following properties
\item{data}{A p x n matrix containing the generated data.}
\item{labels}{A vector giving numerical labels for each data point.
The data points with same label belong in the same segment.}
\item{coords}{A 2 x n matrix with the spatial coordinates.}
\item{center_points}{A vector of length \code{n_segments} containing
the center points for each segment.}
}
\description{
Generates nonstationary Gaussian spatial data with
changing mean and variance by segments. \loadmathjax
}
\details{
The method chooses uniformly \code{n_segments} center points from the provided
spatial locations. The locations are divided into segments based on the
lowest distance to any of the center points. The data is generated from
normal distribution by sampling unique variance from distribution
\mjeqn{Unif(0.1, 5)}{ascii} for each segment. If \code{random_mean = TRUE}
the means for each segment are sampled from distribution \mjeqn{Unif(-5, 5)}{ascii}.
If \code{random_mean = FALSE} the constant zero mean is used for every segment.
}
\examples{
coords <- matrix(runif(2000, 0, 1), ncol = 2)
data_obj <- generate_nonstationary_spatial_data_by_segments(1000, 3, coords, 10)
}
