% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iVAEar_radial.R
\name{iVAEar_radial}
\alias{iVAEar_radial}
\title{iVAEar with Radial Basis Functions for Spatial-Temporal Data}
\usage{
iVAEar_radial(
  data,
  spatial_locations,
  time_points,
  latent_dim,
  n_s,
  elevation = NULL,
  spatial_dim = 2,
  spatial_basis = c(2, 9),
  temporal_basis = c(9, 17, 37),
  ar_order = 1,
  elevation_basis = NULL,
  seasonal_period = NULL,
  max_season = NULL,
  week_component = FALSE,
  spatial_kernel = "gaussian",
  epochs,
  batch_size,
  ...
)
}
\arguments{
\item{data}{A matrix of observed data (n x p), where each row is an observation and each column is a feature.}

\item{spatial_locations}{A matrix of spatial locations corresponding to each observation in \code{data}.}

\item{time_points}{A vector of time points corresponding to each observation in \code{data}.}

\item{latent_dim}{The latent dimension of iVAE.}

\item{n_s}{The number of unique spatial locations in the data.}

\item{elevation}{Optional vector of elevations corresponding to each observation in \code{data}.
Default is \code{NULL}.}

\item{spatial_dim}{An integer specifying the number of spatial dimensions. Default is 2 (e.g., latitude and longitude).}

\item{spatial_basis}{A vector specifying the number of spatial radial basis functions at different resolution levels.}

\item{temporal_basis}{A vector specifying the number of temporal radial basis functions at different resolution levels.}

\item{ar_order}{An autoregressive order used in iVAEar.}

\item{elevation_basis}{Optional vector specifying the number of elevation radial basis functions at different resolution levels. 
Default is \code{NULL}. Has to be provided if \code{elevation} is provided.}

\item{seasonal_period}{Optional integer specifying a seasonal period, if applicable (e.g., 12 for monthly data with a 
yearly period). Default is \code{NULL}.}

\item{max_season}{An integer giving the maximum number of seasonal periods considered. 
Can be used in case the forecasting period have a season number which is not present in the training data.}

\item{week_component}{A boolean value indicating if daily changes within a week is considered. 
Can be used only if the data is observed daily in time.}

\item{spatial_kernel}{A string specifying the kernel to use for spatial data. 
The options are \code{"gaussian"} and \code{"wendland"} Default is \code{"gaussian"}.}

\item{epochs}{Integer specifying the number of training epochs for the iVAE model.}

\item{batch_size}{Integer specifying the batch size for training the iVAE model.}

\item{...}{Additional arguments passed to the underlying \code{\link{iVAEar}} function.}
}
\value{
A fitted iVAEar object of class \code{iVAEradial_st}, which inherits from class \code{\link{iVAEar}}.
In addition the object has the following fields:
  \item{min_coords, max_coords}{Minimum and maximum coordinates of the spatial locations.}
  \item{min_time_point, max_time_point}{Minimum and maximum time points.}
  \item{min_elevation, max_elevation}{Minimum and maximum elevation values, if \code{elevation} is provided.}
  \item{spatial_basis}{The radial basis function configuration for spatial data.}
  \item{temporal_basis}{The radial basis function configuration for temporal data.}
  \item{elevation_basis}{The radial basis function configuration for elevation data, if applicable.}
  \item{spatial_kernel}{The kernel used for spatial data.}
  \item{seasonal_period}{The seasonal period if provided.}
}
\description{
Fits an identifiable Variational Autoencoder with an autoregressive prior (iVAEar) to spatio-temporal data. 
The method constructs the auxiliary data based on spatial and temporal radial basis functions.
}
\details{
This function applies radial basis functions (RBF) to the spatial, temporal, and possibly elevation data to create 
auxiliary variables. These auxiliary variables are passed into the iVAE model to capture dependencies in the latent space 
across space, time, and elevation.

For more details of forming radial basis function based, see \insertCite{sipila2024modelling}{NonlinearBSS}
}
\examples{
p <- 3
n_time <- 100
n_spat <- 50
coords_time <- cbind(
    rep(runif(n_spat), n_time), rep(runif(n_spat), n_time),
    rep(1:n_time, each = n_spat)
)
data_obj <- generate_nonstationary_spatio_temporal_data_by_segments(
    n_time,
    n_spat, p, 5, 10, coords_time
)
latent_data <- data_obj$data
# Generate artificial observed data by applying a nonlinear mixture
obs_data <- mix_data(latent_data, 2)

# Increase the number of epochs for better performance.
resiVAE <- iVAEar_radial(
  data = obs_data, 
  spatial_locations = coords_time[, 1:2],
  time_points = coords_time[, 3],
  latent_dim = p, 
  n_s = n_spat,
  epochs = 1,
  batch_size = 64
)
cormat <- cor(resiVAE$IC, latent_data)
cormat
absolute_mean_correlation(cormat)

}
\seealso{
\code{\link{iVAEar}}
\code{\link{iVAEar_segmentation}}
}
